+++ Models/ElectroFence/ElectroFenceStatusDto.cs
@@ -22,8 +22,8 @@ namespace Dess.Api.Models.ElectroFence
     public bool HvChargeFault { get; set; }
     public bool HvDischargeFault { get; set; }
 
-    public double HvVoltage { get; set; }
-    public double Temperature { get; set; }
+    public short HvVoltage { get; set; }
+    public short Temperature { get; set; }
     public BatteryStatus BatteryStatus { get; set; }
     public byte BatteryLevel { get; set; }
